requirements.txt
pandas
plotly.graph_objects
plotly.subplots